backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
Django==3.2.25
django-celery==3.1.17
django-json-widget==2.1.0
django-simple-history==3.3.0
gunicorn==23.0.0
idna==3.10
importlib-metadata<5.0
kombu==5.2.4
packaging==24.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.0.8
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
vine==5.1.0
wcwidth==0.2.14
zipp==3.15.0
